Conflicts:
	src/System.Text.RegularExpressions/tests/project.lock.json
	src/System.Threading.Overlapped/tests/project.lock.json
	src/System.Threading.Tasks.Dataflow/tests/project.lock.json
	src/System.Threading.Tasks.Parallel/tests/project.lock.json
	src/System.Threading.Tasks/tests/project.lock.json
	src/System.Threading.Timer/tests/project.lock.json
	src/System.Threading/tests/project.lock.json
	src/System.Xml.ReaderWriter/tests/Readers/CharCheckingReader/project.lock.json
	src/System.Xml.ReaderWriter/tests/Readers/CustomReader/project.lock.json
	src/System.Xml.ReaderWriter/tests/Readers/FactoryReader/project.lock.json
	src/System.Xml.ReaderWriter/tests/Readers/NameTable/project.lock.json
	src/System.Xml.ReaderWriter/tests/Readers/ReaderSettings/project.lock.json
	src/System.Xml.ReaderWriter/tests/Readers/SubtreeReader/project.lock.json
	src/System.Xml.ReaderWriter/tests/Readers/WrappedReader/project.lock.json
	src/System.Xml.ReaderWriter/tests/Writers/RwFactory/project.lock.json
	src/System.Xml.ReaderWriter/tests/Writers/XmlWriterApi/project.lock.json
	src/System.Xml.ReaderWriter/tests/XmlConvert/project.lock.json
	src/System.Xml.ReaderWriter/tests/XmlReader/ReadContentAs/project.lock.json
	src/System.Xml.ReaderWriter/tests/XmlReader/Tests/project.lock.json
	src/System.Xml.ReaderWriter/tests/XmlReader/XmlResolver/project.lock.json
	src/System.Xml.ReaderWriter/tests/XmlReaderLib/project.lock.json
	src/System.Xml.ReaderWriter/tests/XmlWriter/project.lock.json
	src/System.Xml.XDocument/tests/Properties/project.lock.json
	src/System.Xml.XDocument/tests/SDMSample/project.lock.json
	src/System.Xml.XDocument/tests/Streaming/project.lock.json
	src/System.Xml.XDocument/tests/TreeManipulation/project.lock.json
	src/System.Xml.XDocument/tests/XDocument.Common/project.lock.json
	src/System.Xml.XDocument/tests/XDocument.Test.ModuleCore/project.lock.json
	src/System.Xml.XDocument/tests/axes/project.lock.json
	src/System.Xml.XDocument/tests/events/project.lock.json
	src/System.Xml.XDocument/tests/misc/project.lock.json
	src/System.Xml.XDocument/tests/xNodeBuilder/project.lock.json
	src/System.Xml.XDocument/tests/xNodeReader/project.lock.json
	src/System.Xml.XPath.XDocument/tests/project.lock.json
	src/System.Xml.XPath.XmlDocument/tests/project.lock.json
	src/System.Xml.XPath/tests/project.lock.json
	src/System.Xml.XmlDocument/tests/project.lock.json
	src/System.Xml.XmlSerializer/tests/project.lock.json

+++ src/System.IO.MemoryMappedFiles/src/System/IO/MemoryMappedFiles/MemoryMappedView.cs
@@ -80,8 +80,8 @@ namespace System.IO.MemoryMappedFiles
         /// <param name="extraMemNeeded">The amount <paramref name="newSize"/> and <paramref name="newOffset"/> were shifted.</param>
         /// <param name="newOffset">The shifted offset based on the <paramref name="allocationGranularity"/>.</param>
         private static void ValidateSizeAndOffset(
-            long size, long offset, int allocationGranularity,
-            out ulong newSize, out ulong extraMemNeeded, out ulong newOffset)
+            long size, long offset, long allocationGranularity,
+            out ulong newSize, out long extraMemNeeded, out long newOffset)
         {
             Debug.Assert(size >= 0);
             Debug.Assert(offset >= 0);
@@ -90,9 +90,9 @@ namespace System.IO.MemoryMappedFiles
             // Determine how much extra memory needs to be allocated to align on the size of allocationGranularity.
             // The newOffset is then moved down by that amount, and the newSize is increased by that amount.
 
-            extraMemNeeded = (ulong)offset % (ulong)allocationGranularity;
-            newOffset = (ulong)offset - extraMemNeeded;
-            newSize = (size != MemoryMappedFile.DefaultSize) ? (ulong)size + extraMemNeeded : 0;
+            extraMemNeeded = offset % allocationGranularity;
+            newOffset = offset - extraMemNeeded;
+            newSize = (size != MemoryMappedFile.DefaultSize) ? (ulong)size + (ulong)extraMemNeeded : 0;
 
             if (IntPtr.Size == 4 && newSize > uint.MaxValue)
             {
