+++ src/System.Linq.Expressions/src/System/Linq/Expressions/Interpreter/AddInstruction.cs
@@ -55,7 +55,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (Int16)unchecked((Int16)l + (Int16)r);
+                    frame.Data[frame.StackIndex - 2] = unchecked((Int16)((Int16)l + (Int16)r));
                 }
                 frame.StackIndex--;
                 return +1;
@@ -74,7 +74,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (Int64)unchecked((Int64)l + (Int64)r);
+                    frame.Data[frame.StackIndex - 2] = unchecked((Int64)((Int64)l + (Int64)r));
                 }
                 frame.StackIndex--;
                 return +1;
@@ -93,7 +93,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (UInt16)unchecked((UInt16)l + (UInt16)r);
+                    frame.Data[frame.StackIndex - 2] = unchecked((UInt16)((UInt16)l + (UInt16)r));
                 }
                 frame.StackIndex--;
                 return +1;
@@ -112,7 +112,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (UInt32)unchecked((UInt32)l + (UInt32)r);
+                    frame.Data[frame.StackIndex - 2] = unchecked((UInt32)((UInt32)l + (UInt32)r));
                 }
                 frame.StackIndex--;
                 return +1;
@@ -131,7 +131,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (UInt64)unchecked((UInt64)l + (UInt64)r);
+                    frame.Data[frame.StackIndex - 2] = unchecked((UInt64)((UInt64)l + (UInt64)r));
                 }
                 frame.StackIndex--;
                 return +1;
@@ -248,7 +248,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (Int16)checked((Int16)l + (Int16)r);
+                    frame.Data[frame.StackIndex - 2] = checked((Int16)((Int16)l + (Int16)r));
                 }
                 frame.StackIndex--;
                 return +1;
@@ -267,7 +267,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (Int64)checked((Int64)l + (Int64)r);
+                    frame.Data[frame.StackIndex - 2] = checked((Int64)((Int64)l + (Int64)r));
                 }
                 frame.StackIndex--;
                 return +1;
@@ -286,7 +286,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (UInt16)checked((UInt16)l + (UInt16)r);
+                    frame.Data[frame.StackIndex - 2] = checked((UInt16)((UInt16)l + (UInt16)r));
                 }
                 frame.StackIndex--;
                 return +1;
@@ -305,7 +305,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (UInt32)checked((UInt32)l + (UInt32)r);
+                    frame.Data[frame.StackIndex - 2] = checked((UInt32)((UInt32)l + (UInt32)r));
                 }
                 frame.StackIndex--;
                 return +1;
@@ -324,7 +324,7 @@ namespace System.Linq.Expressions.Interpreter
                 }
                 else
                 {
-                    frame.Data[frame.StackIndex - 2] = (UInt64)checked((Int16)l + (Int16)r);
+                    frame.Data[frame.StackIndex - 2] = checked((UInt64)((UInt64)l + (UInt64)r));
                 }
                 frame.StackIndex--;
                 return +1;
